libreria 1.4

+++ tp10/portABD.c
@@ -18,22 +18,22 @@ typedef struct
 
 typedef struct
 {
-    uint_least8_t b0    :1;
-    uint_least8_t b1    :1;
-    uint_least8_t b2    :1;
-    uint_least8_t b3    :1;       
-    uint_least8_t b4    :1;
-    uint_least8_t b5    :1;
-    uint_least8_t b6    :1;
-    uint_least8_t b7    :1;
-    uint_least8_t b8    :1;
-    uint_least8_t b9    :1;
-    uint_least8_t b10   :1;
-    uint_least8_t b11   :1;       
-    uint_least8_t b12   :1;
-    uint_least8_t b13   :1;
-    uint_least8_t b14   :1;
-    uint_least8_t b15   :1;
+    uint_least16_t b0    :1;
+    uint_least16_t b1    :1;
+    uint_least16_t b2    :1;
+    uint_least16_t b3    :1;       
+    uint_least16_t b4    :1;
+    uint_least16_t b5    :1;
+    uint_least16_t b6    :1;
+    uint_least16_t b7    :1;
+    uint_least16_t b8    :1;
+    uint_least16_t b9    :1;
+    uint_least16_t b10   :1;
+    uint_least16_t b11   :1;       
+    uint_least16_t b12   :1;
+    uint_least16_t b13   :1;
+    uint_least16_t b14   :1;
+    uint_least16_t b15   :1;
     
 }bit16_t;
